learning/proto/requirements.txt

#
# Required for PAM SOUND
#

pydub
simpleaudio

#
# Required for OLED screen
#

adafruit-blinka
luma.oled
pillow